fixed organization of code
took api keys out of the code, deleted up redundant endpoints, endpoints should now all be found in the rewards users teams and challenges blueprints in flask-server
