+++ moosement_backend/config.py
@@ -1,0 +1,3 @@
+import os
+
+MONGO_URI = os.getenv("MONGO_URI")
