+++ moosement_backend/flask-server/app.py
@@ -4,6 +4,7 @@ from flask_cors import CORS
 from challenges import challenges_bp  # Import the challenges blueprint
 from users import users_bp  # Example of another blueprint for user-related routes
 from teams import teams_bp  # Example of another blueprint for team-related routes
+from rewards import rewards_bp
 
 def create_app():
     """
@@ -15,6 +16,8 @@ def create_app():
 
     # Register the Blueprint for challenges
     app.register_blueprint(challenges_bp, url_prefix='/api/challenges')
+
+    app.register_blueprint(rewards_bp, url_prefix='/api/rewards')
     
     # Register the Blueprint for users (if you have one)
     app.register_blueprint(users_bp, url_prefix='/api/users')
